requirements.txt
python-binance==1.0.32
python-dotenv
